requirements.txt
pyaudio =  0.2.11
speechrecognition =  3.10.4
pydub =   0.25.1
pyannote.audio =  3.3.1
librosa = 0.10.2.post1
tensorflow = 2.16.2